test: design tokens resolution for ssr
